+++ controllers/event_record_router.py
@@ -1,7 +1,13 @@
 from fastapi import APIRouter, HTTPException
 import traceback
 
-from domain.dtos.event_record_dto import EliminationEventRecordDto, EventRecordDto, QuarteredEventRecordDto, RaceEventRecordDto, SprintEventRecordDto
+from domain.dtos.event_record_dto import (
+    EliminationEventRecordDto,
+    EventRecordDto,
+    QuarteredEventRecordDto,
+    RaceEventRecordDto,
+    SprintEventRecordDto,
+)
 from domain.event_record_services.event_record_service import get_event_records
 from domain.event_service import get_event_by_id
 from domain.exceptions.event_not_found_exception import EventNotFoundException
@@ -11,30 +17,36 @@ router = APIRouter(prefix="/event-records", tags=["event-records"])
 
 
 @router.get("/quartered", response_model=list[QuarteredEventRecordDto])
-def get_quartered(event_id: int, playback_tick: int = None) -> list[QuarteredEventRecordDto]:
-    """ Get quartered event records by actions. """
+def get_quartered(
+    event_id: int, playback_tick: int = None
+) -> list[QuarteredEventRecordDto]:
+    """Get quartered event records by actions."""
     return _get_event_records(event_id, playback_tick=playback_tick)
 
 
 @router.get("/endurance", response_model=list[RaceEventRecordDto])
-def get_endurance(event_id: int, is_playback: bool = False, playback_tick: int = None) -> list[RaceEventRecordDto]:
-    """ Get endurance race event records by actions. """
-    return _get_event_records(event_id, is_playback, playback_tick=playback_tick)
-
-
-@router.get("/sprint", response_model=list[SprintEventRecordDto])
-def get_sprint(event_id: int, is_playback: bool = False, playback_tick: int = None) -> list[SprintEventRecordDto]:
-    """ Get sprint race event records by actions. """
-    return _get_event_records(event_id, is_playback, playback_tick=playback_tick)
-
-
-@router.get("/elimination", response_model=list[EliminationEventRecordDto])
-def get_elimination(event_id: int, playback_tick: int = None) -> list[EliminationEventRecordDto]:
-    """ Get elimination event records by actions. """
+def get_endurance(event_id: int, playback_tick: int = None) -> list[RaceEventRecordDto]:
+    """Get endurance race event records by actions."""
     return _get_event_records(event_id, playback_tick=playback_tick)
 
 
-def _get_event_records(event_id: int, is_playback: bool = False, playback_tick: int = None) -> list[EventRecordDto]:
+@router.get("/sprint", response_model=list[SprintEventRecordDto])
+def get_sprint(event_id: int, playback_tick: int = None) -> list[SprintEventRecordDto]:
+    """Get sprint race event records by actions."""
+    return _get_event_records(event_id, playback_tick=playback_tick)
+
+
+@router.get("/elimination", response_model=list[EliminationEventRecordDto])
+def get_elimination(
+    event_id: int, playback_tick: int = None
+) -> list[EliminationEventRecordDto]:
+    """Get elimination event records by actions."""
+    return _get_event_records(event_id, playback_tick=playback_tick)
+
+
+def _get_event_records(
+    event_id: int, playback_tick: int = None
+) -> list[EventRecordDto]:
     event = None
     try:
         event = get_event_by_id(event_id)
@@ -42,13 +54,20 @@ def _get_event_records(event_id: int, is_playback: bool = False, playback_tick: 
         raise HTTPException(status_code=404, detail="EVENT_NOT_FOUND")
     except Exception as e:
         traceback.print_exc()
-        raise HTTPException(status_code=500, detail=f"Error querying event: {e.with_traceback(None)}")
+        raise HTTPException(
+            status_code=500, detail=f"Error querying event: {e.with_traceback(None)}"
+        )
 
     if event is None:
         raise HTTPException(status_code=404, detail="EVENT_NOT_FOUND")
 
     try:
-        return get_event_records(event.actions, event.competitors, event.type, is_playback, playback_tick)
+        return get_event_records(
+            event.actions, event.competitors, event.type, playback_tick
+        )
     except Exception as e:
         traceback.print_exc()
-        raise HTTPException(status_code=500, detail=f"Error getting event records: {e.with_traceback(None)}")
+        raise HTTPException(
+            status_code=500,
+            detail=f"Error getting event records: {e.with_traceback(None)}",
+        )